refactor: change structure map
